backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
motor==3.6.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12
